파일백업

+++ backup/install/library/apr.sh
@@ -17,8 +17,10 @@ set -e
 # 사용 하는 extglob 쉘 옵션 shopt 내장 명령을 사용 하 여 같은 확장된 패턴 일치 연산자를 사용
 shopt -s extglob
 
-rm -rf ${SERVER_HOME}/${APR_ALIAS}
-rm -rf ${SERVER_HOME}/${PROGRAME_HOME}/${APR_HOME}
+if [[ ! -z ${APR_ALIAS} ]]; then
+    rm -rf ${SERVER_HOME}/${APR_ALIAS}
+fi
+rm -rf ${SERVER_HOME}${PROGRAME_HOME}/${APR_HOME}
 
 printf "\e[00;32m| ${APR_HOME} / ${APR_ICONV_NAME} / ${APR_UTIL_HOME} install start...\e[00m\n"
 
@@ -49,7 +51,7 @@ fi
 tar xvzf ${APR_NAME}
 cd ${SRC_HOME}/${APR_HOME}
 
-./configure --prefix=${SERVER_HOME}/${PROGRAME_HOME}/${APR_HOME}
+./configure --prefix=${SERVER_HOME}${PROGRAME_HOME}/${APR_HOME}
 make
 make install
 sleep 0.5
@@ -67,8 +69,8 @@ fi
 tar xvzf ${APR_UTIL_NAME}
 cd ${SRC_HOME}/${APR_UTIL_HOME}
 
-./configure --prefix=${SERVER_HOME}/${PROGRAME_HOME}/${APR_HOME}                \
-            --with-apr=${SERVER_HOME}/${PROGRAME_HOME}/${APR_HOME}
+./configure --prefix=${SERVER_HOME}${PROGRAME_HOME}/${APR_HOME}                 \
+            --with-apr=${SERVER_HOME}${PROGRAME_HOME}/${APR_HOME}
 make
 make install
 
@@ -85,15 +87,17 @@ fi
 tar xvzf ${APR_ICONV_NAME}
 cd ${SRC_HOME}/${APR_ICONV_HOME}
 
-./configure --prefix=${SERVER_HOME}/${PROGRAME_HOME}/${APR_HOME}                \
-            --with-apr=${SERVER_HOME}/${PROGRAME_HOME}/${APR_HOME}
+./configure --prefix=${SERVER_HOME}${PROGRAME_HOME}/${APR_HOME}                 \
+            --with-apr=${SERVER_HOME}${PROGRAME_HOME}/${APR_HOME}
 make
 make install
 sleep 0.5
 
 # ----------------------------------------------------------------------------------------------------------------------
-cd ${SERVER_HOME}
-ln -s ./${PROGRAME_HOME}/${APR_HOME} ${APR_ALIAS}
+if [[ ! -z ${APR_ALIAS} ]]; then
+    cd ${SERVER_HOME}
+    ln -s ./${PROGRAME_HOME}/${APR_HOME} ${APR_ALIAS}
+fi
 
 # ----------------------------------------------------------------------------------------------------------------------
 # Install source delete
